Use ropes instead of concatenating strings.
So far implemented in HTML writer only.

The advantage is not speed -- this is actually
slower than the old version -- but flexibility.

A rope (here) is an array containing strings,
functions, and ropes.  At the end of parsing, the
rope is walked and a string constructed by concatenating
its elements in order.

The functions are evaluated
after parsing, and can access globals constructed
during parsing, such as a list of references or footnotes.
So we no longer need to do two passes for things like this.
This will be really useful when we add things like
citations.

As a bonus, reference link definitions now work even
in embedded contexts like block quotes.