Added 'standalone' target to Makefile.
This creates a standalone executable version of lunamark in the
standalone directory.  It doesn't require that lua or other
lua libraries be installed.  The lua interpreter is baked into
the executable.

It also uses a version of lpeg that has been modified to allow
for larger grammars.

See the patch in standalone/.